+++ ConverterApp2/app/src/main/java/com/example/converterapp2/myModel.kt
@@ -1,5 +1,4 @@
 package com.example.converterapp2
 
-data class MyModel(
-        var resize : Double=0.0) {
+data class MyModel(var resize : Int = 0) {
 }
